requirements.txt
torch
torchvision
pillow
scikit-learn
matplotlib
numpy
